// src/app/models/effort.model.ts
//  -------------------------------------------------------------------
//  effort.model.ts
//  -------------------------------------------------------------------

export class Effort {
  k_EffortId: number;
  effortCode: string;
  effortName: string;
  description: string
  sysCreateDateTime: Date;
  sysCreateUser: string;
  sysCreateMethod: string;
  sysRecordStatus: string;
  sysUpdateDateTime: Date;
  sysUpdateUser: string;
  sysUpdateMethod: string;

}