// src/app/models/priority.model.ts
//  -------------------------------------------------------------------
//  priority.model.ts
//  -------------------------------------------------------------------

export class Priority {
  k_PriorityId: number;
  priorityCode: string;
  priorityName: string;
  description: string
  sysCreateDateTime: Date;
  sysCreateUser: string;
  sysCreateMethod: string;
  sysRecordStatus: string;
  sysUpdateDateTime: Date;
  sysUpdateUser: string;
  sysUpdateMethod: string;

}