// src/app/models/project.model.ts
//  -------------------------------------------------------------------
//  project.model.ts
//  -------------------------------------------------------------------

export class Project {
  k_ProjectId: number;
  projectCode: string;
  projectName: string;
  description: string
  sysCreateDateTime: Date;
  sysCreateUser: string;
  sysCreateMethod: string;
  sysRecordStatus: string;
  sysUpdateDateTime: Date;
  sysUpdateUser: string;
  sysUpdateMethod: string;

}