// src/app/models/issueType.model.ts
//  -------------------------------------------------------------------
//  issueSType.model.ts
//  -------------------------------------------------------------------

export class IssueType {
  k_IssueTypeId: number;
  issueTypeCode: string;
  issueTypeName: string;
  description: string
  sysCreateDateTime: Date;
  sysCreateUser: string;
  sysCreateMethod: string;
  sysRecordStatus: string;
  sysUpdateDateTime: Date;
  sysUpdateUser: string;
  sysUpdateMethod: string;

}