// src/app/models/component.model.ts
//  -------------------------------------------------------------------
//  component.model.ts
//  -------------------------------------------------------------------

export class MyComponent {
  k_ComponentId: number;
  componentCode: string;
  componentName: string;
  description: string
  sysCreateDateTime: Date;
  sysCreateUser: string;
  sysCreateMethod: string;
  sysRecordStatus: string;
  sysUpdateDateTime: Date;
  sysUpdateUser: string;
  sysUpdateMethod: string;

}