// src/app/models/viewIssue.model.ts
//  -------------------------------------------------------------------
//  viewIssue.model.ts
//  -------------------------------------------------------------------


export class ViewIssue {
    k_IssueId: number;
    projectCode: string;
    issueNumber: number;
    issueTypeName: string;
    issueStatusName: string;
    releaseNo: string;
    priorityCode: string;
    effortCode: string;
    assignedUser: string;
    uRL: string;
    issueName: string;
    description: string;
    component1Name: string;
    level1Name: string;
    component2Name: string;
    level2Name: string;
    component3Name: string;
    level3Name: string;
    component4Name: string;
    level4Name: string;
    component5Name: string;
    level5Name: string;
    recordDate: Date;
    recordDateStr: string;
    assignDate: Date;
    assignDateStr: string;
    targetDate: Date;
    targetDateStr: string;
    doneDate: Date;
    doneDateStr: string;
    testDate: Date;
    testDateStr: string;
    completeDate: Date;
    completeDateStr: string;
    gotoLink: string;

}